producer/requirements.txt
requests
kafka-python
faker